TR-39-refactor: 컴포넌트 폴더 정리
