+++ src/components/AppLayout/AppLayout.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Outlet } from "react-router-dom"
-import GlobalHeader from "./GlobalHeader"
+import GlobalHeader from "../GlobalHeader/GlobalHeader"
 
 const AppLayout = () => {
   return (
