+++ src/components/GlobalHeader/GlobalHeader.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link, useLocation } from "react-router-dom"
-import Button from "./Button"
+import Button from "../Button/Button"
 import logo from "../assets/logo.png"
 import styles from "./GlobalHeader.module.css"
